requirements.txt
asgiref==3.3.4
Django==3.2.3
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.2.0
Markdown==3.3.4
pdfkit==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
python-slugify==5.0.2
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
zipp==3.4.1
